requirements.txt
gradio==4.16.0
openai==1.14.3
gtts==2.5.1
python-dotenv==1.0.1
SpeechRecognition==3.10.1